pyproject.toml
[project]
name = "projetos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.7",
    "djangorestframework>=3.16.1",
]
[tool.uv.scripts]
run = "manage.py runserver"
migrate = "manage.py migrate"
makemig = "manage.py makemigrations"
superuser = "manage.py createsuperuser"
